refactor: consolidate crash reporting test fixtures and harden assertions
- Move RecordingGameCrashDiagnostics and RecordingSensitiveDataSink to
  shared TestData/RecordingCrashSinks.cs
- Remove duplicate private fixture classes from BaseGameTests and
  CrashContextPublisherTests
- Rename tests for clearer intent and add DTX path redaction assertion
